+++ crates/turbopack-ecmascript/src/tree_shake/util.rs
@@ -9,11 +9,14 @@ use swc_core::{
             ArrowExpr, AssignPatProp, AssignTarget, ClassDecl, ClassExpr, Constructor, DefaultDecl,
             ExportDefaultDecl, ExportNamedSpecifier, ExportSpecifier, Expr, FnDecl, FnExpr,
             Function, Id, Ident, ImportSpecifier, MemberExpr, MemberProp, NamedExport, Param, Pat,
-            Prop, PropName, VarDeclarator,
+            Prop, PropName, VarDeclarator, *,
         },
         visit::{noop_visit_type, Visit, VisitWith},
     },
 };
+use turbo_tasks::RcStr;
+
+use crate::TURBOPACK_HELPER;
 
 #[derive(Debug, Default, Clone, Copy)]
 enum Mode {
@@ -379,4 +382,129 @@ where
     };
     n.visit_with(&mut v);
     v.bindings
+}
+
+pub fn should_skip_tree_shaking(m: &Program, special_exports: &[RcStr]) -> bool {
+    if let Program::Module(m) = m {
+        for item in m.body.iter() {
+            match item {
+                // Skip turbopack helpers.
+                ModuleItem::ModuleDecl(ModuleDecl::Import(ImportDecl {
+                    with, specifiers, ..
+                })) => {
+                    if let Some(with) = with.as_deref().and_then(|v| v.as_import_with()) {
+                        for item in with.values.iter() {
+                            if item.key.sym == *TURBOPACK_HELPER {
+                                // Skip tree shaking if the import is from turbopack-helper
+                                return true;
+                            }
+                        }
+                    }
+
+                    // TODO(PACK-3150): Tree shaking has a bug related to ModuleExportName::Str
+                    for s in specifiers.iter() {
+                        if let ImportSpecifier::Named(is) = s {
+                            if matches!(is.imported, Some(ModuleExportName::Str(..))) {
+                                return true;
+                            }
+                        }
+                    }
+                }
+
+                // Tree shaking has a bug related to ModuleExportName::Str
+                ModuleItem::ModuleDecl(ModuleDecl::ExportNamed(NamedExport {
+                    src: Some(..),
+                    specifiers,
+                    ..
+                })) => {
+                    for s in specifiers {
+                        if let ExportSpecifier::Named(es) = s {
+                            if matches!(es.orig, ModuleExportName::Str(..))
+                                || matches!(es.exported, Some(ModuleExportName::Str(..)))
+                            {
+                                return true;
+                            }
+                        }
+                    }
+                }
+
+                // Skip sever actions
+                ModuleItem::Stmt(Stmt::Expr(ExprStmt {
+                    expr: box Expr::Lit(Lit::Str(Str { value, .. })),
+                    ..
+                })) => {
+                    if value == "use server" {
+                        return true;
+                    }
+                }
+
+                // Skip special reexports that are recognized by next.js
+                ModuleItem::ModuleDecl(ModuleDecl::ExportDecl(ExportDecl {
+                    decl: Decl::Var(box VarDecl { decls, .. }),
+                    ..
+                })) => {
+                    for decl in decls {
+                        if let Pat::Ident(name) = &decl.name {
+                            if special_exports.iter().any(|s| **s == *name.sym) {
+                                return true;
+                            }
+                        }
+                    }
+                }
+
+                // Skip special reexports that are recognized by next.js
+                ModuleItem::ModuleDecl(ModuleDecl::ExportDecl(ExportDecl {
+                    decl: Decl::Fn(f),
+                    ..
+                })) => {
+                    if special_exports.iter().any(|s| **s == *f.ident.sym) {
+                        return true;
+                    }
+                }
+
+                _ => {}
+            }
+        }
+
+        let mut visitor = UseServerFinder::default();
+        m.visit_with(&mut visitor);
+        if visitor.abort {
+            return true;
+        }
+
+        for item in m.body.iter() {
+            if item.is_module_decl() {
+                return false;
+            }
+        }
+    }
+
+    true
+}
+
+#[derive(Default)]
+struct UseServerFinder {
+    abort: bool,
+}
+
+impl Visit for UseServerFinder {
+    fn visit_expr_stmt(&mut self, e: &ExprStmt) {
+        e.visit_children_with(self);
+
+        if let Expr::Lit(Lit::Str(Str { value, .. })) = &*e.expr {
+            if value == "use server" {
+                self.abort = true;
+            }
+        }
+    }
+
+    fn visit_stmt(&mut self, n: &Stmt) {
+        if self.abort {
+            return;
+        }
+
+        n.visit_children_with(self);
+    }
+
+    noop_visit_type!();
 }
